+++ app/src/main/java/com/example/anibey_codex_tfg/ui/navigation/Screen.kt
@@ -6,6 +6,5 @@ sealed interface Screen {
     @Serializable
     data object Welcome : Screen
 
-    @Serializable
-    data object Login : Screen
+    @Serializable data class Login(val isRegister: Boolean) : Screen
 }